hacky fix to a tiny problem
When different background colors are set, the title text may be invisible or difficult to read. One way of solving this is by flipping the color of the title between white and black based on the contrast of the background. A more lazy fix is to use mix blend mode, like here but hey atleast it works 😛